refactor: 컨트롤러에 있던 비지니스 로직을 서비스로 이관
